requirements.txt
mysql-connector-python
psycopg2-binary
typing-extensions
colorama
python-dotenv